# services/user/sdk/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/asset/{assetIdentity}": {
            "get": {
                "operationId": "AssetController_getAsset",
                "parameters": [
                    {
                        "name": "assetIdentity",
                        "required": true,
                        "in": "path",
                        "description": "Asset identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetAssetResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            },
            "delete": {
                "operationId": "AssetController_deleteAsset",
                "parameters": [
                    {
                        "name": "assetIdentity",
                        "required": true,
                        "in": "path",
                        "description": "Asset identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
        "/api/v1/assetGroup/{assetGroupIdentity}": {
            "get": {
                "operationId": "AssetController_getAssetGroup",
                "parameters": [
                    {
                        "name": "assetGroupIdentity",
                        "required": true,
                        "in": "path",
                        "description": "Asset group identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetAssetGroupResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            },
            "delete": {
                "operationId": "AssetController_deleteAssetGroup",
                "parameters": [
                    {
                        "name": "assetGroupIdentity",
                        "required": true,
                        "in": "path",
                        "description": "Asset group identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
        "/api/v1/user/{userIdentity}/asset": {
            "get": {
                "operationId": "AssetController_getUserAssets",
                "parameters": [
                    {
                        "name": "userIdentity",
                        "required": true,
                        "in": "path",
                        "description": "User identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetUserAssetsResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
        "/api/v1/user/{userIdentity}/assetGroup": {
            "get": {
                "operationId": "AssetController_getUserAssetGroups",
                "parameters": [
                    {
                        "name": "userIdentity",
                        "required": true,
                        "in": "path",
                        "description": "User identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetUserAssetGroupsResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            },
            "put": {
                "operationId": "AssetController_createAssetGroupForUser",
                "parameters": [
                    {
                        "name": "userIdentity",
                        "required": true,
                        "in": "path",
                        "description": "User identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAssetGroupBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateAssetGroupForUserResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "DuplicateEntryError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
        "/api/v1/user/{userIdentity}/asset/stock": {
            "put": {
                "operationId": "AssetController_createStockAssetForUser",
                "parameters": [
                    {
                        "name": "userIdentity",
                        "required": true,
                        "in": "path",
                        "description": "User identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStockAssetBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateStockAssetForUserResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "DuplicateEntryError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
        "/api/v1/user/{userIdentity}/asset/realEstate": {
            "put": {
                "operationId": "AssetController_createRealEstateAssetForUser",
                "parameters": [
                    {
                        "name": "userIdentity",
                        "required": true,
                        "in": "path",
                        "description": "User identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRealEstateAssetBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateRealEstateAssetForUserResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "DuplicateEntryError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
        "/api/v1/assetGroup/{assetGroupIdentity}/asset/stock": {
            "put": {
                "operationId": "AssetController_createStockAssetForGroup",
                "parameters": [
                    {
                        "name": "assetGroupIdentity",
                        "required": true,
                        "in": "path",
                        "description": "Asset group identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStockAssetBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateStockAssetForAssetGroupResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "DuplicateEntryError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "Finance API",
        "description": "Finance API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "scopes": {
                            "openid": "openid",
                            "profile": "profile",
                            "email": "email"
                        },
                        "authorizationUrl": "https://dev-tzti25ivp5zmox2s.us.auth0.com/authorize"
                    }
                },
                "in": "header",
                "name": "Authorization",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
...
            "HttpErrorResponse": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "integer",
                        "description": "The http code returned by the server"
                    },
                    "message": {
                        "type": "string",
                        "description": "A message describing the error that occurred"
                    }
                },
                "required": [
                    "statusCode",
                    "message"
                ]
            },
...
            "SuccessResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "success"
                ]
            },
...
            "GetAssetResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AssetResponse"
                    }
                },
                "required": [
                    "success",
                    "data"
                ]
            },
...
            "GetAssetGroupResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/GetAssetGroupResponseData"
                    }
                },
                "required": [
                    "success",
                    "data"
                ]
            },
            "GetUserAssetsResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PaginatedAssetResponse"
                    }
                },
                "required": [
                    "success",
                    "data"
                ]
            },
...
            "GetUserAssetGroupsResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PaginatedAssetGroupResponse"
                    }
                },
                "required": [
                    "success",
                    "data"
                ]
            },
...
            "CreateStockAssetForUserResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AssetResponse"
                    }
                },
                "required": [
                    "success",
                    "data"
                ]
            },
...
            "CreateAssetGroupBody": {
                "type": "object",
                "properties": {}
            },
            "CreateAssetGroupForUserResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AssetGroupResponse"
                    }
                },
                "required": [
                    "success",
                    "data"
                ]
            },
...
        }
    }
}